requirements.txt
streamlit>=1.28.0
streamlit-webrtc>=0.47.0
streamlit-autorefresh>=1.0.0
av>=10.0.0
numpy>=1.24.0
scipy>=1.10.0
python-dotenv>=1.0.0
assemblyai>=0.42.0
